# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Libreria",
        "description": "Esta API puede usarse para obtener información sobre los libros, autores y generos de la librería, y crear un pedido.\n *      Si la url para la API no tiene el formato adecuado o no existe, se retornará un json con un mensaje indicándoselo.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://pixel-pioneers-laravel-git-etapa2-pixel-pioneer.vercel.app/rest/v1"
        }
    ],
    "paths": {
...
        "/autores/{id}": {
            "get": {
                "tags": [
                    "Autores"
                ],
                "description": "Devuelve la información del autor con la id especificada. Incluyendo los libros de su autoría que esten disponibles en la tienda.",
                "operationId": "81fd6d7c19b8f8dbd15266322f1e82e9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Autor encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/AutorShow"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe Autor con la id provista",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Autor No Encontrado"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/libros/{id}": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "description": "Devuelve la información del libro con la id especificada",
                "operationId": "e95837776da37090df39d0d80ee6511d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libro encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/LibroShow"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No existe libro con la id provista",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "Libro No Encontrado"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/{titulo}/searchTitle": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "description": "Devuelve los libros que contengan en su titulo el termino especificado",
                "operationId": "346f9e4a88457feb0f17fca4384a99b0",
                "parameters": [
                    {
                        "name": "titulo",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libros encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LibroIndex"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron libros con el termino especificado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "No se encontraron libros con el termino especificado"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/{genero}/searchGenre": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "description": "Devuelve los libros que contengan el termino especificado en el parametro",
                "operationId": "4a56da8964d52fdaf327882c11128106",
                "parameters": [
                    {
                        "name": "genero",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libros encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LibroIndex"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron libros con el genero especificado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "No se encontraron libros con el genero especificado"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/libros/{nombreAutor}/searchAuthor": {
            "get": {
                "tags": [
                    "Libros"
                ],
                "description": "Devuelve los libros escritos por el autor especificado. El parametro puede contener mas de un nombre o apellido siempre y cuando esten separados por espacios.",
                "operationId": "43911efbde908d23122c5def2f76cf1c",
                "parameters": [
                    {
                        "name": "nombreAutor",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Libros encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/LibroIndex"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No se encontraron libros con los autores especificado",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "error": "No se encontraron libros con los autores especificado"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AutorShow": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Nombre Autor"
                    },
                    "apellido": {
                        "type": "string",
                        "example": "Apellido Autor"
                    },
                    "libros": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LibroIndex"
                        }
                    }
                },
                "type": "object"
            },
...
            "LibroShow": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Titulo Libro"
                    },
                    "precio": {
                        "type": "string",
                        "example": "940338"
                    },
                    "cantidad_paginas": {
                        "type": "integer",
                        "example": 1177
                    },
                    "disponibilidad": {
                        "type": "boolean",
                        "example": false
                    },
                    "url_imagen": {
                        "type": "string",
                        "example": "https://pzwiki.net/w/images/a/ac/SkillBookAnim_120px.gif"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripción del libro"
                    },
                    "autores": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "154"
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Mathilde"
                                },
                                "apellido": {
                                    "type": "string",
                                    "example": "Reinger"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "generos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "6"
                                },
                                "nombre_genero": {
                                    "type": "string",
                                    "example": "Terror"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "LibroIndex": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "titulo": {
                        "type": "string",
                        "example": "Titulo Libro"
                    },
                    "precio": {
                        "type": "string",
                        "example": "940338"
                    },
                    "cantidad_paginas": {
                        "type": "integer",
                        "example": 1177
                    },
                    "disponibilidad": {
                        "type": "boolean",
                        "example": false
                    },
                    "url_imagen": {
                        "type": "string",
                        "example": "https://pzwiki.net/w/images/a/ac/SkillBookAnim_120px.gif"
                    },
                    "autores": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "154"
                                },
                                "nombre": {
                                    "type": "string",
                                    "example": "Mathilde"
                                },
                                "apellido": {
                                    "type": "string",
                                    "example": "Reinger"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "generos": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": "6"
                                },
                                "nombre_genero": {
                                    "type": "string",
                                    "example": "Terror"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
...
}